refactored share-button
