+++ lib/src/Components/Map/Subcomponents/ItemPopupComponents/HeaderView/hooks.ts
@@ -35,10 +35,7 @@ export const useNavigationUrl = (coordinates?: [number, number]) => {
 export const useShareLogic = (item?: Item) => {
   const shareUrl = window.location.href
   const shareTitle = item?.name ?? 'Utopia Map Item'
-  const inviteLink =
-    item?.secrets && item.secrets.length > 0
-      ? `${window.location.origin}/invite/${item.secrets[0].secret}`
-      : shareUrl
+  const inviteLink = shareUrl
 
   const copyLink = () => {
     navigator.clipboard
