+++ lib/src/Components/Map/Subcomponents/ItemPopupComponents/StartEndView.tsx
@@ -8,25 +8,19 @@ import type { Item } from '#types/Item'
  */
 export const StartEndView = ({ item }: { item?: Item }) => {
   return (
-    <div className='tw:flex tw:flex-row tw:mb-4 tw:mt-4 tw:bg-base-300 tw:px-4 tw:py-3 tw:rounded-selector tw:w-full'>
-      <div className='tw:basis-2/5 tw:flex tw:flex-row tw:font-bold'>
-        <CalendarIcon className='tw:h-4 tw:w-4 tw:mr-2' />
-        <time
-          className='tw:align-middle'
-          dateTime={item && item.start ? item.start.substring(0, 10) : ''}
-        >
+    <div className='tw:flex tw:flex-row tw:mb-2.5 tw:mt-2.5 tw:bg-base-200 tw:px-3 tw:py-2.5 tw:rounded-selector tw:w-full'>
+      <div className='tw:basis-2/5 tw:flex tw:flex-row tw:items-center tw:font-bold'>
+        <CalendarIcon className='tw:h-5 tw:w-5 tw:mr-2' />
+        <time dateTime={item && item.start ? item.start.substring(0, 10) : ''}>
           {item && item.start ? new Date(item.start).toLocaleDateString() : ''}
         </time>
       </div>
-      <div className='tw:basis-1/5 tw:place-content-center'>
+      <div className='tw:basis-1/5 tw:flex tw:items-center tw:justify-center'>
         <span>-</span>
       </div>
-      <div className='tw:basis-2/5 tw:flex tw:flex-row tw:font-bold'>
-        <CalendarIcon className='tw:h-4 tw:w-4 tw:mr-2' />
-        <time
-          className='tw:align-middle'
-          dateTime={item && item.end ? item.end.substring(0, 10) : ''}
-        >
+      <div className='tw:basis-2/5 tw:flex tw:flex-row tw:items-center tw:font-bold'>
+        <CalendarIcon className='tw:h-5 tw:w-5 tw:mr-2' />
+        <time dateTime={item && item.end ? item.end.substring(0, 10) : ''}>
           {item && item.end ? new Date(item.end).toLocaleDateString() : ''}
         </time>
       </div>
